+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased] 
+### Changed
+- Read a register instead of doing a ping in the Ethernet's NetStatusListener
+
 ## [7.0.1] - 2023-04-04
 ### Fixed
 - Recover old Monitoring/Disturbance compatibility