+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased] 
+### Changed
+- Read a register instead of doing a ping in the Ethernet's NetStatusListener
+- Use inspect instead of pkg_resources to find the path to the FoE application.
+- Call FoE application with utf-8 encoding.
+
+
 ## [7.0.1] - 2023-04-04
 ### Fixed
 - Recover old Monitoring/Disturbance compatibility